refactor(resize): extract detectEdge into a dependency-free module
The unit test imported detectEdge from useNodeResize.ts, which transitively
pulls in appStore and the electron-log + xterm graph. Under the node test
environment that graph leaves a handle open, so the test process never exits
and the CI "Unit tests" step hangs indefinitely.

Move detectEdge, getCursorForEdge, the ResizeEdge type and the thresholds into
resizeEdge.ts (no store/hook/electron imports) and re-export them from
useNodeResize.ts so existing callers are unchanged. The test now imports the
pure module, so it tests the same function without loading any browser code.
